+++ src/tests/StreamTasksTests.java
@@ -2,45 +2,48 @@ package tests;
 
 import java.util.List;
 import java.util.Set;
+import java.util.function.Supplier;
 import java.util.stream.Stream;
 
+import tests.presets.IntNumbers;
 import tests.presets.Names;
 import tests.utils.Test;
 import tests.utils.TestInputCollection;
 import tests.utils.TestInputConstant;
+import tests.utils.TestInputSupplier;
 import tests.utils.TestResult;
 
 public abstract class StreamTasksTests {
     
-    @Test (order = 1)
+    @Test (order = 0)
     @TestResult (wrap = List.class)
     public abstract Set <String> task1 (
         @TestInputCollection (presets = {Names.class}, constant = {1, 5}, variation = 5)
         List <String> names
     );
     
-    @Test (order = 2)
+    @Test (order = 0)
     @TestResult (wrap = Set.class)
     public abstract List <String> task2 (
         @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 1.1}, variation = 10)
         List <String> names
     );
     
-    @Test (order = 3)
+    @Test (order = 0)
     @TestResult (wrap = Set.class)
     public abstract List <String> task3 (
         @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 1.1}, variation = 10)
         Set <String> names
     );
     
-    @Test (order = 4)
+    @Test (order = 0)
     @TestResult (wrap = Void.class)
     public abstract int task4 (
         @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 1.1}, variation = 10)
         Set <String> names
     );
     
-    @Test (order = 5)
+    @Test (order = 0)
     @TestResult (wrap = Void.class)
     public abstract int task5 (
         @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 1.1}, variation = 10)
@@ -49,14 +52,14 @@ public abstract class StreamTasksTests {
         int length
     );
     
-    @Test (order = 6)
+    @Test (order = 0)
     @TestResult (wrap = Void.class)
     public abstract List <String> task6 (
         @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 2.1})
         Stream <String> names
     );
     
-    @Test (order = 6)
+    @Test (order = 0)
     @TestResult (wrap = List.class)
     public abstract Stream <String> task7 (
         @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 2.1})
@@ -64,5 +67,36 @@ public abstract class StreamTasksTests {
         @TestInputCollection (presets = {Names.class}, percentage = {0.33, 0.45})
         Stream <String> names2
     );
+    
+    @Test (order = 0)
+    @TestResult (wrap = List.class)
+    public abstract Stream <String> task8 (
+        @TestInputCollection (presets = {Names.class}, percentage = {1.01, 1.05, 2.1})
+        Stream <String> names,
+        @TestInputCollection (presets = {Names.class}, percentage = {0.33, 0.45})
+        Stream <String> names2,
+        @TestInputConstant (sequence = {0}, parameter = 0, variation = 10)
+        int limit
+    );
+    
+    @Test (order = 0)
+    @TestResult (wrap = Void.class)
+    public abstract <T> List <T> task9 (
+        @TestInputCollection (presets = {Names.class, IntNumbers.class}, percentage = {0.91, 1.05, 1.51})
+        Stream <T> values
+    );
+    
+    @Test (order = 1)
+    @TestResult (wrap = Void.class)
+    public abstract <T> List <T> task10 (
+        @TestInputSupplier (presets = {Names.class, IntNumbers.class})
+        Supplier <T> generator, 
+        @TestInputConstant (constant = {100}, variation = 10)
+        int limit
+    );
+    
+    @Test (order = 1)
+    @TestResult (wrap = Void.class)
+    public abstract List <Integer> task11 ();
     
 }
