+++ src/tests/presets/SupplierMode.java
@@ -1,0 +1,8 @@
+package tests.presets;
+
+
+public enum SupplierMode {
+    
+    SEQUENTIAL, SHUFFLED_SEQUENTIAL, RANDOM
+    
+}
